+++ pyccel/ast/numpy_wrapper.py
@@ -9,8 +9,10 @@ Handling the transitions between Python code and C code using (Numpy/C Api).
 
 import numpy as np
 
+from pyccel.utilities.metaclasses import Singleton
+
 from .datatypes         import (NativeInteger, NativeFloat, NativeComplex,
-                                NativeBool, NativeGeneric, NativeVoid)
+                                NativeBool, NativeGeneric, NativeVoid, DataType)
 
 from .cwrapper          import PyccelPyObject
 
@@ -29,6 +31,8 @@ from ..errors.messages import PYCCEL_RESTRICTION_TODO
 errors = Errors()
 
 __all__ = (
+    #--------- DATATYPES ---------
+    'PyccelPyArrayObject',
     #------- CAST FUNCTIONS ------
     'pyarray_to_ndarray',
     #-------CHECK FUNCTIONS ------
@@ -39,9 +43,20 @@ __all__ = (
     'array_get_data',
     'array_get_c_step',
     'array_get_f_step',
+    'PyArray_SetBaseObject',
     #-------OTHERS--------
-    'get_numpy_max_acceptable_version_file'
+    'get_numpy_max_acceptable_version_file',
 )
+
+class PyccelPyArrayObject(DataType, metaclass=Singleton):
+    """
+    Datatype representing a `PyArrayObject`.
+
+    Datatype representing a `PyArrayObject` which is the
+    class used to hold NumPy array objects in Python.
+    """
+    __slots__ = ()
+    _name = 'PyArrayObject'
 
 #-------------------------------------------------------------------
 #                      Numpy functions
@@ -124,6 +139,14 @@ array_get_data  = FunctionDef(name   = 'nd_data',
                            body      = [],
                            arguments = [FunctionDefArgument(Variable(dtype=NativeVoid(), name = 'o', is_optional=True))],
                            results   = FunctionDefResult(Variable(dtype=NativeVoid(), name = 'v', memory_handling='alias', rank = 1, class_type = NativeVoid())))
+
+PyArray_SetBaseObject = FunctionDef(name   = 'PyArray_SetBaseObject',
+                                    body      = [],
+                                    arguments = [FunctionDefArgument(Variable(dtype=PyccelPyArrayObject(), name = 'arr', memory_handling='alias')),
+                                                 FunctionDefArgument(Variable(dtype=PyccelPyObject(), name = 'obj', memory_handling='alias'))],
+                                    results   = [FunctionDefResult(Variable(dtype=NativeInteger(), name = 'd'))])
+
+import_array = FunctionDef('import_array', (), (), ())
 
 # Basic Array Flags
 # https://numpy.org/doc/stable/reference/c-api/array.html#c.NPY_ARRAY_OWNDATA