+++ tests/epyccel/modules/return_pointers.py
@@ -1,0 +1,6 @@
+# pylint: disable=missing-function-docstring, missing-module-docstring
+
+def return_ambiguous_pointer_to_argument(x: 'int[:]'):
+    y = x
+    w = y
+    return w